# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Filmes",
        "description": "API desenvolvida para uma plataforma de filmes",
        "contact": {
            "email": "josebrunocampanholi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de testes"
        }
    ],
    "paths": {
        "/api/atores": {
            "get": {
                "tags": [
                    "atores"
                ],
                "summary": "Atores cadastrados no sistema",
                "description": "Retorna os atores cadastrados no sistema paginados em 15",
                "operationId": "indexAtor",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "atores": {
                                            "$ref": "#/components/schemas/Ator"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "atores"
                ],
                "summary": "Cria um ator",
                "description": "Método para criar um novo ator no sistema",
                "operationId": "storeAtor",
                "requestBody": {
                    "description": "Obrigatório nome do autor",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "bruno"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Novo Ator cadastrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro nos parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Informe todos os dados obrigatórios"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/categorias": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Categorias cadastradas no sistema",
                "description": "Retorna as categorias cadastrados no sistema paginados em 15",
                "operationId": "indexCategoria",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categorias": {
                                            "$ref": "#/components/schemas/Categoria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "categorias"
                ],
                "summary": "Cria uma categoria",
                "description": "Método para criar uma nova categoria no sistema",
                "operationId": "storeAtor",
                "requestBody": {
                    "description": "Obrigatório nome da categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Ação"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Nova Categoria cadastrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro nos parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Informe todos os dados obrigatórios"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Mostra Categoria",
                "description": "Retorna categoria pelo ID da url",
                "operationId": "showCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categoria": {
                                            "$ref": "#/components/schemas/Categoria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "categorias"
                ],
                "summary": "Atualiza Categoria",
                "description": "Atualiza uma Categoria, de acordo com o ID",
                "operationId": "updatecategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Nome da Categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "bruno"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Ator atualizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro nos parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Informe todos os dados obrigatórios"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "categorias"
                ],
                "summary": "Deleta categoria",
                "description": "Deleta Categoria pelo ID da url",
                "operationId": "destroyCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Categoria",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Sem conteúdo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/categorias/{id}/filmes": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Categoria dos Filmes",
                "description": "Retorna todos os filmes que essa categoria tem",
                "operationId": "filmesDaCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "filmes": {
                                            "$ref": "#/components/schemas/Filme"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Ator": {
                "required": [
                    "nome"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome do Ator",
                        "type": "string",
                        "format": "string",
                        "example": "John"
                    },
                    "nacionalidade": {
                        "description": "Nacionalidade do Ator",
                        "type": "string",
                        "format": "string"
                    },
                    "obs": {
                        "description": "Abservações a respeito do Ator",
                        "type": "string",
                        "format": "string"
                    },
                    "created_at": {
                        "description": "Criação inicial timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Última mudança timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
            "Categoria": {
                "required": [
                    "nome"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome da Categoria",
                        "type": "string",
                        "format": "string",
                        "example": "Ação"
                    },
                    "created_at": {
                        "description": "Criação inicial timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Última mudança timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Filme": {
                "required": [
                    "nome"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome do Filme",
                        "type": "string",
                        "format": "string",
                        "example": "Vingadores"
                    },
                    "created_at": {
                        "description": "Criação inicial timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Última mudança timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Coloque um Token no formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}